Treat providers as implementations
